[Bug] Plan a partial target's row gate inside its partition scope; gate rows against a projected aggregate
A `partial datasource ... complete where city = 'A'` target with a row gate
that reads an aggregate (`where id = cluster_id`, in any of its three
spellings) failed with "no complete sources": the claim was ANDed into a
cross-row stage, `drop_excluded_partials` stood down for the whole stage, and
the city partials had to exhaust the entire `city` enum. The claim's scalar
atoms now lead as their own stage whenever the gate computes across rows, so
the union is proven complete over the target's own slice and the gate is
evaluated after the union and aggregation. Purely scalar gates keep the
single-stage AND and plan byte-identically.

The same prune then exposed a long-standing silent leak: `where id = anchor
select id, anchor` (or a coalesce/flag derived from the aggregate) returned
rows violating the WHERE. The atom is hosted at the select-scope aggregate as
a pre-aggregation filter and the ROOT stream fans the surviving groups back out
through the FINAL join. Four sites conspired to hide it: the FINAL re-filter
guard only tripped when the ROOT alone exposed every input; a bare key peer of
a by-aggregate contributed nothing to a BASIC's grain; the terminal-projection
rule read a grain key the expression uses as a grain-only passenger; and an
ancestor aggregate's pre-filter was summarised as preexisting for every
descendant. Each is corrected at its source; the FINAL re-filter trips only
when an uncovered contributor alone supplies an output the host grain does not
functionally determine, which keeps the tpc-ds/tpc-h corpus byte-identical.

DuckDB and the base dialect now reverse-map `double`, `float8` and `double
precision` to DataType.DOUBLE instead of FLOAT.

+++ trilogy/dialect/duckdb.py
@@ -469,9 +469,9 @@ class DuckDBDialect(BaseDialect):
         "timestamptz": DataType.TIMESTAMP,
         "float4": DataType.FLOAT,
         "real": DataType.FLOAT,
-        "double": DataType.FLOAT,
-        "float8": DataType.FLOAT,
-        "double precision": DataType.FLOAT,
+        "double": DataType.DOUBLE,
+        "float8": DataType.DOUBLE,
+        "double precision": DataType.DOUBLE,
         "decimal": DataType.NUMERIC,
     }
